requirements.txt
aiohttp~=3.8.4
python-dotenv~=0.21.1
web3~=5.31.3
